Fix interpunction missing from messages
