+++ New Unity Project/Assets/Script/DataObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dc7dd3a309bf8f245ac793559d708856
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
