+++ New Unity Project/Assets/Script/DataObjects/Message.cs
@@ -1,0 +1,7 @@
+[System.Serializable]
+public class Message
+{
+    public string date;
+    public string author;
+    public string text;
+}
