+++ New Unity Project/Assets/Script/DataObjects/Message.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8458d1090351de40a42d73e8d11e729
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
