+++ New Unity Project/Assets/Script/DataObjects/Post.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99548ef0a3a98ee4baef52917456a97c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
