+++ New Unity Project/Assets/Script/DataObjects/PostContainer.cs
@@ -1,0 +1,8 @@
+﻿namespace Assets.Script.DataObjects
+{
+    [System.Serializable]
+    public class PostContainer
+    {
+        public Post[] posts;
+    }
+}
