+++ New Unity Project/Assets/Script/DataObjects/PostContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 727637dcd6b67c54baf0b649ddca07d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
