+++ New Unity Project/Assets/Script/Phone.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+public class Phone : MonoBehaviour
+{
+    public Color LinkColor;
+
+}
