+++ New Unity Project/Assets/Script/Phone.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c94bf4e6937f6b049829cdaa47fbf618
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
